Dependencies (#889)
* chore(deps): update dependency @astrojs/rss to v2.4.4

* chore(deps): update dependency @playwright/test to v1.37.0

* chore(deps): update dependency prettier-plugin-tailwindcss to v0.5.1

* chore(deps): update dependency sass to v1.65.1

* chore(deps): update typescript-eslint monorepo to v6.3.0

* chore(deps): update dependency astro to v2.10.5

* Update dependencies and streamline YAML styling

The changes in this commit update various dependencies as noted in the diff, such as "astro-netlify-cms", "astro", and several "@astrojs" packages. The version updates ensure the project dependencies are up-to-date, potentially incorporating bug fixes, security patches, and feature enhancements from those packages.

In addition to the dependency

* Update dependencies and streamline YAML styling

The changes in this commit update various dependencies as noted in the diff, such as "astro-netlify-cms", "astro", and several "@astrojs" packages. The version updates ensure the project dependencies are up-to-date, potentially incorporating bug fixes, security patches, and feature enhancements from those packages.

In addition

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>