chore(deps): Update sharp version in pnpm-lock.yaml file
This commit updates the version of the sharp. This was required to match the current version of other dependencies. This change will help ensure proper functioning of the project which relies on this latest version of sharp for image handling. Additionally, this commit also tidies up resolution hash representation for cleaner reading in the pnpm lock file.