Update various dependencies in the pnpm-lock.yaml
This commit updates the versions of several dependencies specified in the pnpm-lock.yaml file to their latest versions for improved features and security patches. This includes updating the versions for '@astrojs/tailwind', '@astrojs/vue', '@typescript-eslint/eslint-plugin', and 'sass' among others. This update helps to ensure that our project's dependencies are up-to-date and no critical updates are missed for stable and secure project builds.